pasan todos los tests !!!

+++ salaDeReuniones_TPI/src/auxiliar.h
@@ -12,7 +12,7 @@ bool freqValida(int freq);
 bool conPrecision(tiempo t,int p);
 bool enRango(audio a, int prof);
 bool sonTodosCeros(vector<int> s);
-tiempo enSegundos(int n,int freq);
+tiempo enSegundos(float n, float freq);
 tiempo duracion(audio a,int freq);
 bool duraMasDe(tiempo t, audio a, int freq);
 vector<int> subSeq(vector<int> s,int i, int j);
